+++ openai/plugin/src/examples/kotlin/addressbook/env/startUserPlugin.kt
@@ -8,8 +8,8 @@ import addressbook.user.UserPluginSettings.PORT
 import org.http4k.cloudnative.env.Environment.Companion.ENV
 import org.http4k.connect.openai.auth.OpenAIPluginId
 import org.http4k.connect.openai.model.Email
-import org.http4k.connect.openai.plugins.PluginIntegrationBuilder
-import org.http4k.connect.openai.plugins.UserPluginIntegrationBuilder
+import org.http4k.connect.openai.plugins.PluginIntegration
+import org.http4k.connect.openai.plugins.UserPluginIntegration
 import org.http4k.core.Uri
 import org.http4k.core.then
 import org.http4k.core.with
@@ -20,7 +20,7 @@ import org.http4k.filter.debug
 import org.http4k.server.SunHttp
 import org.http4k.server.asServer
 
-fun startUserPlugin(): PluginIntegrationBuilder {
+fun startUserPlugin(): PluginIntegration {
     val env = ENV.with(
         PORT of 30000,
         PLUGIN_BASE_URL of Uri.of("http://localhost:30000"),
@@ -33,7 +33,7 @@ fun startUserPlugin(): PluginIntegrationBuilder {
         .asServer(SunHttp(PORT(env)))
         .start()
 
-    return UserPluginIntegrationBuilder(
+    return UserPluginIntegration(
         BasicAuth("") { UserDirectory().auth(it) != null },
         OpenAIPluginId.of("userplugin"),
         PLUGIN_BASE_URL(env)