+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
-### v3.41.1.0 (uncut)
-- **http4k-connect-openai** - More tweaks to the openai plugin model
+### v3.42.0.0
+- **http4k-connect-openai-plugin** - [New module!] OpenAI plugin development SDK. http4k-connect provides APIs to create plugins for all 3 plugin authorization types - User, Service and OAuth.
+- **http4k-connect-openai-fake** - Plugins can now be installed into the `FakeOpenAI` server. All 3 plugin auth types
+are supported. 
 
 ### v3.41.0.0
 - **http4k-connect-*** - Upgrade dependencies.